log, style improvements, translation

+++ src/lib/helper.js
@@ -1,6 +1,12 @@
+// export function dateToString(date) {
+// 	const date_obj = new Date(date);
+// 	const options = { weekday: 'long', year: 'numeric', month: 'short', day: 'numeric', hour: 'numeric', minute: 'numeric' };
+// 	return date_obj.toLocaleDateString('de-DE', options);
+// }
+
 export function dateToString(date) {
 	const date_obj = new Date(date);
-	const options = { weekday: 'long', year: 'numeric', month: 'short', day: 'numeric', hour: 'numeric', minute: 'numeric' };
+	const options = { weekday: 'long', year: 'numeric', month: 'numeric', day: 'numeric', hour: 'numeric', minute: 'numeric' };
 	return date_obj.toLocaleDateString('de-DE', options);
 }
